chore: moves test to CreateReceivables.test.tsx
